Moves spawnWithSpinner to utils
