+++ lib/commands/create.js
@@ -6,8 +6,6 @@ const util = require('../util');
 const config = require('../../config');
 const fs = require('fs');
 const { stripIndent } = require('common-tags');
-const ora = require('ora');
-const { spawn } = require('child_process');
 
 module.exports = function (program) {
   program
@@ -68,7 +66,10 @@ module.exports = function (program) {
       }
 
       try {
-        await spanWithSpinner();
+        await util.spawnWithSpinner('npm install', {
+          spinnerMessage: 'Installing packages. This will take a little while...',
+          codeMessage: 'Install spawn process error with code'
+        });
       } catch (error) {
         await util.error('create', 'Error installing packages');
         /* eslint-disable-next-line no-console */
@@ -109,37 +110,4 @@ module.exports = function (program) {
 
 function replaceInConfig(regex, replacement) {
   util.replaceInFiles([ './app.js', './package.json' ], regex, replacement);
-}
-
-async function spanWithSpinner () {
-  const spinner = ora({
-    text: 'Installing packages...',
-    // stream: process.stdout,
-    interval: 100,
-    isEnabled: true
-  });
-
-  return new Promise((resolve, reject) => {
-    try {
-      const installer = spawn('npm install', {
-        shell: true
-      });
-
-      spinner.start();
-
-      installer.on('exit', (code) => {
-        spinner.stop();
-
-        if (code !== null && code !== 0) {
-          reject(new Error(`Install spawn process error with code ${code}`));
-          return;
-        }
-
-        resolve();
-      });
-    } catch (error) {
-      reject(error);
-    }
-  });
-
 }
