+++ lib/util.js
@@ -9,6 +9,8 @@ const cliVersion = require('../package.json').version;
 const confUtils = require('./conf-utils');
 const packageJson = require('package-json');
 const semver = require('semver');
+const ora = require('ora');
+const { spawn } = require('child_process');
 
 module.exports = util;
 
@@ -143,6 +145,43 @@ util.secret = function() {
     string += s;
   }
   return string;
+};
+
+util.spawnWithSpinner = async function (command, options = {
+  spinnerMessage: 'Processing...',
+  codeMessage: 'Spawn process error code:'
+}) {
+  if (!command) {
+    throw new Error('No command provided to spawnWithSpinner');
+  }
+
+  const spinner = ora({
+    text: options.spinnerMessage,
+    interval: 100,
+    isEnabled: true
+  });
+
+  return new Promise((resolve, reject) => {
+    try {
+      const installer = spawn(command, {
+        shell: true
+      });
+
+      spinner.start();
+
+      installer.on('exit', (code) => {
+        spinner.stop();
+
+        if (code !== null && code !== 0) {
+          return reject(new Error(`${options.codeMessage} ${code}`));
+        }
+
+        return resolve();
+      });
+    } catch (error) {
+      return reject(error);
+    }
+  });
 };
 
 async function checkIfUpdated () {
